fix: refactor folder tree
