+++ src/components/Layout/Layout.tsx
@@ -1,7 +1,7 @@
 import { Component, ReactNode } from 'react';
 import { Outlet } from 'react-router-dom';
-import Footer from './Footer';
-import Header from './Header';
+import Footer from '../Footer/Footer';
+import Header from '../Header/Header';
 
 class Layout extends Component {
   render(): ReactNode {
